Changed quarkus version to 999-snapshot to fix issue with injection in tests Refactored repositories to PanacheRepositoryBase Removed

+++ src/main/java/org/springframework/samples/petclinic/service/ClinicService.java
@@ -33,36 +33,36 @@ import org.springframework.samples.petclinic.model.Visit;
  */
 public interface ClinicService {
 
-	Pet findPetById(long id) ;
+	Pet findPetById(Integer id) ;
 	Collection<Pet> findAllPets() ;
 	void savePet(Pet pet) ;
 	void deletePet(Pet pet) ;
 
-	Collection<Visit> findVisitsByPetId(long petId);
-	Visit findVisitById(long visitId) ;
+	Collection<Visit> findVisitsByPetId(Integer petId);
+	Visit findVisitById(Integer visitId) ;
 	Collection<Visit> findAllVisits() ;
 	void saveVisit(Visit visit) ;
 	void deleteVisit(Visit visit) ;
 	
-	Vet findVetById(long id) ;
+	Vet findVetById(Integer id) ;
 	Collection<Vet> findVets() ;
 	Collection<Vet> findAllVets() ;
 	void saveVet(Vet vet) ;
 	void deleteVet(Vet vet) ;
 	
-	Owner findOwnerById(long id) ;
+	Owner findOwnerById(Integer id) ;
 	Collection<Owner> findAllOwners() ;
 	void saveOwner(Owner owner) ;
 	void deleteOwner(Owner owner) ;
 	Collection<Owner> findOwnerByLastName(String lastName) ;
 
-	PetType findPetTypeById(long petTypeId);
+	PetType findPetTypeById(Integer petTypeId);
 	Collection<PetType> findAllPetTypes() ;
 	Collection<PetType> findPetTypes() ;
 	void savePetType(PetType petType) ;
 	void deletePetType(PetType petType) ;
 	
-	Specialty findSpecialtyById(long specialtyId);
+	Specialty findSpecialtyById(Integer specialtyId);
 	Collection<Specialty> findAllSpecialties() ;
 	void saveSpecialty(Specialty specialty) ;
 	void deleteSpecialty(Specialty specialty) ;
